pyproject.toml
[tool.black]
line-length = 110
target-version = ["py39"]

[tool.isort]
line_length = 110
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
default_section = "THIRDPARTY"

[tool.poetry]
name = "mars_rover"
version = "0.1.0"
description = ""
authors = ["Tomas Menito <tomasmenito@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.dev-dependencies]
pytest-randomly = "^3.10.1"
pytest-cov = "^2.12.1"
Faker = "^8.12.1"
pytest = "^6.2.5"
factory-boy = "^3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
